Adapt the slides according the extraordinary general assembly
Include the amount of the budget for the corporate identity, to make it
more explicit

Determine who is going to decide on the procedure regarding the
corporate identity, to make it more explicit

Adapt the links to the changes of the bylaws (since there was a rebase
of the branch).

Add a link to the sponsor contract template.